Some more warnings removed.

+++ src/Bpp/Numeric/Matrix/MatrixTools.h
@@ -1102,11 +1102,11 @@ namespace bpp
             if (uMin < uSubMin) {
               // put in current k, and go back to that k.
               // continue augmenting path i - j1 with i0.
-              free[--k] = i0; 
+              free[--k] = static_cast<size_t>(i0); 
             } else { 
               // no further augmenting reduction possible.
               // store i0 in list of free rows for next phase.
-              free[numFree++] = i0; 
+              free[numFree++] = static_cast<size_t>(i0); 
             }
           }
         }
